feat: update reward points system to use float values; add reward points history feature

+++ backend/models/reward_points.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Date, DateTime, Text, ForeignKey
+from sqlalchemy import Column, Integer, Float, Date, DateTime, Text, ForeignKey
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
@@ -13,7 +13,7 @@ class RewardPointRedemption(Base):
     user_id = Column(UUID(as_uuid=True), ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True)
     account_id = Column(UUID(as_uuid=True), ForeignKey("accounts.id", ondelete="CASCADE"), nullable=False, index=True)
     date = Column(Date, nullable=False)
-    points_used = Column(Integer, nullable=False)
+    points_used = Column(Float, nullable=False)
     description = Column(Text, nullable=True)
     created_at = Column(DateTime(timezone=True), server_default=func.current_timestamp(), nullable=False)
     updated_at = Column(DateTime(timezone=True), server_default=func.current_timestamp(), onupdate=func.current_timestamp(), nullable=False)
